move zones panel sections into dedicated sections folder
