+++ src/components/panels/zones/ZonesPanel.tsx
@@ -1,8 +1,9 @@
 import "../advanced/AdvancedPanel.css";
+import "./ZonesPanel.css";
 import type { Settings } from "../../../store";
-import FailsafeSection from "./FailsafeSection";
-import CustomStopZoneSection from "./CustomStopZoneSection";
-import ProcessListSection from "./ProcessListSection";
+import FailsafeSection from "./sections/FailsafeSection";
+import CustomStopZoneSection from "./sections/CustomStopZoneSection";
+import ProcessListSection from "./sections/ProcessListSection";
 
 // TODO: This still entirely needs to get done. I've just moved everything over here. If you are a back end dev, please be careful changing UI things xD
 // TODO: Custom Stop zones should be like Sequence clicking where you can add as many as you want in a list.
